Implements a "smart" thread pool system
The thread pool can be configured to grow and shrink in size to meet
minimum deadlines. For example, if a task cannot be executed within the
deadline of 10 ms, then the thread pool needs to employ more workers.

Moreover, if workers are idle for too long they will be fired from the
pool.

However it is also possible to simply use a static number of workers, if
desired. This can be configured in "Constants".